+++ patches/src/main/kotlin/app/crimera/patches/instagram/misc/ghostMode/Fingerprint.kt
@@ -1,11 +1,7 @@
 /*
  * Copyright (C) 2026 piko <https://github.com/crimera/piko>
  *
- * This file is part of piko.
- *
- * Any modifications, derivatives, or substantial rewrites of this file
- * must retain this copyright notice and the piko attribution
- * in the source code and version control history.
+ * See the included NOTICE file for GPLv3 §7(b) terms that apply to this code.
  */
 
 package app.crimera.patches.instagram.misc.ghostMode
@@ -17,7 +13,8 @@ import com.android.tools.smali.dexlib2.AccessFlags
 internal object DMSeenFingerprint : Fingerprint(
     returnType = "V",
     accessFlags = listOf(AccessFlags.PUBLIC, AccessFlags.STATIC, AccessFlags.FINAL),
-    filters = listOf(
-        string("mark_thread_seen-")
-    ),
+    filters =
+        listOf(
+            string("mark_thread_seen-"),
+        ),
 )