+++ patches/src/main/kotlin/app/crimera/patches/twitter/entity/Fingerprints.kt
@@ -1,11 +1,7 @@
 /*
  * Copyright (C) 2026 piko <https://github.com/crimera/piko>
  *
- * This file is part of piko.
- *
- * Any modifications, derivatives, or substantial rewrites of this file
- * must retain this copyright notice and the piko attribution 
- * in the source code and version control history.
+ * See the included NOTICE file for GPLv3 §7(b) terms that apply to this code.
  */
 
 package app.crimera.patches.twitter.entity
@@ -62,21 +58,24 @@ internal object TweetShortTextFingerprint : Fingerprint(
     name = "getShortText",
 )
 
-internal object TweetActionsHandlerFingerprint: Fingerprint(
-   strings = listOf("content_author")
+internal object TweetActionsHandlerFingerprint : Fingerprint(
+    strings = listOf("content_author"),
 )
 
-internal class TweetOriginalNameFingerprint(definingClass: String) : Fingerprint(
-    accessFlags = listOf(AccessFlags.PUBLIC, AccessFlags.FINAL),
-    returnType = "Ljava/lang/String;",
-    parameters = emptyList(),
-    definingClass = definingClass,
-    filters = listOf(
-        methodCall(definingClass = "this", returnType = "Ljava/lang/String;"),
-        methodCall(definingClass = "this", returnType = "Ljava/lang/String;"),
-        opcode(Opcode.RETURN_OBJECT)
-    ),
-)
+internal class TweetOriginalNameFingerprint(
+    definingClass: String,
+) : Fingerprint(
+        accessFlags = listOf(AccessFlags.PUBLIC, AccessFlags.FINAL),
+        returnType = "Ljava/lang/String;",
+        parameters = emptyList(),
+        definingClass = definingClass,
+        filters =
+            listOf(
+                methodCall(definingClass = "this", returnType = "Ljava/lang/String;"),
+                methodCall(definingClass = "this", returnType = "Ljava/lang/String;"),
+                opcode(Opcode.RETURN_OBJECT),
+            ),
+    )
 
 internal object TweetMediaEntityClassFingerprint : Fingerprint(
     strings = listOf("EntityList{mEntities="),
@@ -96,9 +95,10 @@ internal object QuotedViewSetAccessibilityFingerprint : Fingerprint(
 )
 
 // --------------- Extended Media Entity
-internal object MediaOptionSheetMediaListVideoDownloaderImplDownloadMethodFingerprint : Fingerprint(
+// Also required for download patch.
+object MediaOptionSheetMediaListVideoDownloaderImplDownloadMethodFingerprint : Fingerprint(
     returnType = "Z",
-    strings = listOf("mediaEntity"),
+    strings = listOf("url", "video_download"),
     custom = { _, classDef ->
         classDef.startsWith("Lcom/twitter/tweetview/core/ui/mediaoptionssheet/")
     },